Update Apollo client library to v3.8.1
The Apollo client library was updated from v2.5.11 to v3.8.1 to keep the project up-to-date and leverage new features and bug fixes provided in the newer version. The changes also include adjustments to existing Apollo-related code so that they're compatible with the newer version, notably the package change from "com.apollographql.apollo" to "com.apollographql.apollo3".