Refactor Gradle dependencies to use version catalog
This commit refactors the build.gradle.kts file to take advantage of Gradle's version catalog feature. Instead of hardcoding plugin versions and dependencies directly within the file, these are now maintained in the libs.versions.toml file. This makes it easier to manage project dependencies as they are now centralized in one file. This pattern is efficient and improves the readability and maintainability of the build file.